+++ Pipfile
@@ -12,6 +12,7 @@ pytest-cov = "*"
 
 [packages]
 redis = "*"
+inject = "*"
 
 [requires]
 python_version = "3.5"
